requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
colorama==0.4.4
fastapi==0.75.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
importlib-metadata==4.11.3
Jinja2==3.1.1
MarkupSafe==2.1.1
pydantic==1.9.0
python-dotenv==0.20.0
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
zipp==3.7.0
